refactored to organise code
