+++ src/main/java/Duke.java
@@ -1,15 +1,12 @@
 import CustomExceptions.DukeException;
 import Model_Classes.*;
+import Operations.Parser;
+import Operations.Storage;
+import Operations.TaskList;
+import Operations.Ui;
 
-import java.io.*;
-import java.lang.reflect.Array;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.Scanner;
-import java.util.spi.AbstractResourceBundleProvider;
 
 /**
  * main class of the Duke program
