+++ src/main/java/Operations/Storage.java
@@ -1,6 +1,7 @@
-package Model_Classes;
+package Operations;
 
 import CustomExceptions.DukeException;
+import Model_Classes.*;
 
 import java.io.*;
 import java.text.DateFormat;
